refactor: #193 PR2 evaluate.ts を役割別 evaluators モジュールへ分割
936行に同居していた評価ロジック (material/pst/hand-value/king-safety/
piece-activity/promotion-threat/piece-safety) を evaluators/ 配下の役割別
モジュールへ分離し、evaluate.ts を合成ルート + 後方互換 re-export に整理した。
PR3 のカード/ドロー価値調整を安全に行うための保守性向上が目的。

振る舞いは完全不変 (evaluate-equivalence.test.ts の 1000 局面 byte-level 一致で
担保)。既存 consumer の import 互換も維持。
