+++ src/lib/shogi/ai/evaluators/pst.ts
@@ -1,0 +1,168 @@
+// 配置評価テーブル (Piece-Square Tables) と PST マップ。
+//
+// Issue #193 / PR2: evaluate.ts から分離。全 PST テーブルと PST_MAP を集約する。
+// 値・並び順は分離元から 1 文字も変えていない (evaluate-equivalence.test.ts の
+// 1000 局面 byte-level 一致で担保)。
+//
+// 配置評価テーブル（先手視点、row=0が前線/敵陣奥, row=8が自陣奥）
+
+// 歩兵の配置ボーナス
+export const PAWN_PST: number[][] = [
+  [30, 30, 30, 30, 30, 30, 30, 30, 30],
+  [20, 20, 20, 20, 20, 20, 20, 20, 20],
+  [10, 10, 10, 10, 10, 10, 10, 10, 10],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [-5, -5, -5, -5, -5, -5, -5, -5, -5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 飛車: 中央制圧・前進ボーナス
+export const ROOK_PST: number[][] = [
+  [15, 10, 10, 10, 15, 10, 10, 10, 15],
+  [15, 10, 10, 10, 15, 10, 10, 10, 15],
+  [10,  5,  5,  5, 10,  5,  5,  5, 10],
+  [ 5,  5,  5,  5, 10,  5,  5,  5,  5],
+  [ 5,  5,  5,  5, 10,  5,  5,  5,  5],
+  [ 0,  0,  0,  0,  5,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  5,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 角行: 対角線制圧
+export const BISHOP_PST: number[][] = [
+  [ 0,  5,  0,  5,  0,  5,  0,  5,  0],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 0,  5, 10,  5,  0,  5, 10,  5,  0],
+  [ 5,  5,  5, 10,  5, 10,  5,  5,  5],
+  [ 0,  5,  0,  5, 15,  5,  0,  5,  0],
+  [ 5,  5,  0,  5,  5,  5,  0,  5,  5],
+  [ 0,  5,  0,  5,  0,  5,  0,  5,  0],
+  [ 5,  0,  5,  0,  5,  0,  5,  0,  5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 金将: 玉周辺守備
+export const GOLD_PST: number[][] = [
+  [ 0,  0,  0,  5,  0,  5,  0,  0,  0],
+  [ 0,  5,  5,  5,  5,  5,  5,  5,  0],
+  [ 0,  5, 10,  5,  5,  5, 10,  5,  0],
+  [ 0,  0,  5, 10, 10, 10,  5,  0,  0],
+  [ 0,  0,  5,  5, 10,  5,  5,  0,  0],
+  [ 0,  0,  0,  5,  5,  5,  0,  0,  0],
+  [ 0,  0,  0,  0,  5,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 銀将: 玉周辺守備 + 攻め参加
+export const SILVER_PST: number[][] = [
+  [ 0,  0,  5,  5,  5,  5,  5,  0,  0],
+  [ 0,  5,  5,  5,  5,  5,  5,  5,  0],
+  [ 0,  5, 10,  5,  5,  5, 10,  5,  0],
+  [ 0,  5,  5, 10,  5, 10,  5,  5,  0],
+  [ 0,  5,  5,  5, 10,  5,  5,  5,  0],
+  [ 0,  0,  5,  5,  5,  5,  5,  0,  0],
+  [ 0,  0,  0,  5,  5,  5,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 桂馬: 後段ペナルティ、前進ボーナス
+export const KNIGHT_PST: number[][] = [
+  [  0,  5, 10, 10, 10, 10, 10,  5,  0],
+  [  0,  5, 10, 15, 15, 15, 10,  5,  0],
+  [  0,  5,  5, 10, 10, 10,  5,  5,  0],
+  [  0,  5,  5,  5,  5,  5,  5,  5,  0],
+  [  0,  0,  5,  5,  5,  5,  5,  0,  0],
+  [  0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [-10,-10,-10,-10,-10,-10,-10,-10,-10],
+  [-15,-15,-15,-15,-15,-15,-15,-15,-15],
+  [-20,-20,-20,-20,-20,-20,-20,-20,-20],
+];
+
+// 香車: 前進ボーナス
+export const LANCE_PST: number[][] = [
+  [20, 20, 20, 20, 20, 20, 20, 20, 20],
+  [15, 15, 15, 15, 15, 15, 15, 15, 15],
+  [10, 10, 10, 10, 10, 10, 10, 10, 10],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [-5, -5, -5, -5, -5, -5, -5, -5, -5],
+  [-5, -5, -5, -5, -5, -5, -5, -5, -5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 玉: 囲い位置ボーナス（左右どちらかの端に寄る）
+export const KING_PST: number[][] = [
+  [-30,-40,-40,-50,-50,-50,-40,-40,-30],
+  [-30,-40,-40,-50,-50,-50,-40,-40,-30],
+  [-30,-40,-40,-50,-50,-50,-40,-40,-30],
+  [-30,-40,-40,-50,-50,-50,-40,-40,-30],
+  [-20,-30,-30,-40,-40,-40,-30,-30,-20],
+  [-10,-20,-20,-30,-30,-30,-20,-20,-10],
+  [ 20, 20,  0,  0,  0,  0,  0, 20, 20],
+  [ 20, 30, 10,  0,  0,  0, 10, 30, 20],
+  [ 20, 30, 10,  0,  0,  0, 10, 30, 20],
+];
+
+// 成り駒（金将と同等の動き）のPST - 前進ボーナスを持つ
+export const PROMOTED_MINOR_PST: number[][] = [
+  [10, 10, 10, 10, 15, 10, 10, 10, 10],
+  [ 5,  5,  5,  5, 10,  5,  5,  5,  5],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 0,  0,  0,  0,  5,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 龍王: 飛車の成り - より強力な前進ボーナス
+export const PROMOTED_ROOK_PST: number[][] = [
+  [25, 20, 20, 20, 25, 20, 20, 20, 25],
+  [20, 15, 15, 15, 20, 15, 15, 15, 20],
+  [15, 10, 10, 10, 15, 10, 10, 10, 15],
+  [10,  5,  5,  5, 15,  5,  5,  5, 10],
+  [ 5,  5,  5,  5, 15,  5,  5,  5,  5],
+  [ 0,  0,  0,  0, 10,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  5,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// 龍馬: 角行の成り
+export const PROMOTED_BISHOP_PST: number[][] = [
+  [10, 10, 10, 10, 10, 10, 10, 10, 10],
+  [10, 15, 10, 10, 15, 10, 10, 15, 10],
+  [10, 10, 20, 10, 10, 10, 20, 10, 10],
+  [10, 10, 10, 20, 10, 20, 10, 10, 10],
+  [10, 10, 10, 10, 25, 10, 10, 10, 10],
+  [10, 10, 10, 10, 10, 10, 10, 10, 10],
+  [10, 10, 10, 10, 10, 10, 10, 10, 10],
+  [ 5,  5,  5,  5,  5,  5,  5,  5,  5],
+  [ 0,  0,  0,  0,  0,  0,  0,  0,  0],
+];
+
+// PSTマップ
+export const PST_MAP: Record<string, number[][] | undefined> = {
+  pawn: PAWN_PST,
+  rook: ROOK_PST,
+  bishop: BISHOP_PST,
+  gold: GOLD_PST,
+  silver: SILVER_PST,
+  knight: KNIGHT_PST,
+  lance: LANCE_PST,
+  king: KING_PST,
+  promoted_pawn: PROMOTED_MINOR_PST,
+  promoted_lance: PROMOTED_MINOR_PST,
+  promoted_knight: PROMOTED_MINOR_PST,
+  promoted_silver: PROMOTED_MINOR_PST,
+  promoted_rook: PROMOTED_ROOK_PST,
+  promoted_bishop: PROMOTED_BISHOP_PST,
+};
